New frontend design and Task!
There's a lot of testing for a new form of input for product names. You can see how you use a notebook as well from here

Please text me if you don't understand anything. Keep playing around with the code.

Warmup task:

Add a button that resets all selections (Deletes all selected products from the selected products listbox)

Task:

right now, when you selecte multiple products, the transfer is fine, but duplicates can be made. so if you select a product, then click the button multiple times, it will keep adding it to the second textbox. I want to make it so that if you re-select the same item, the text "You already have selected this product!" to be printed IN THE CONSOLE, not the program.